filtro PBajos si PBanda no se puede

+++ user_config_DEFAULT.py
@@ -5,7 +5,7 @@
 
 CANT_CANALES=36
 #CANT_CANALES=24
-FS=float(20000)
+FS=float(10000)
 PAQ_USB=6000
 #FAKE_FILE=True #para file hay q cambiar el parser
